Split BatteryGlobals into item class & list class
